refac: Extract code to medicated modules
