+++ .flake8
@@ -1,3 +1,4 @@
 [flake8]
 max-line-length = 99
 exclude = .git/*,.tox/*,docs/*,dist/*,build/*
+ignore = E231,W503
