+++ tests/test_transitions.py
@@ -63,7 +63,11 @@ def test_transition_as_decorator_should_call_method_before_activating_state(
 
 
 @pytest.mark.parametrize(
-    "machine_name", ["traffic_light_machine", "reverse_traffic_light_machine", ]
+    "machine_name",
+    [
+        "traffic_light_machine",
+        "reverse_traffic_light_machine",
+    ],
 )
 def test_cycle_transitions(request, machine_name):
     machine_class = request.getfixturevalue(machine_name)
@@ -158,7 +162,14 @@ def test_transitions_to_the_same_estate_as_itself():
 
 
 class TestReverseTransition(object):
-    @pytest.mark.parametrize("initial_state", ["green", "yellow", "red", ])
+    @pytest.mark.parametrize(
+        "initial_state",
+        [
+            "green",
+            "yellow",
+            "red",
+        ],
+    )
     def test_reverse_transition(self, reverse_traffic_light_machine, initial_state):
         machine = reverse_traffic_light_machine(start_value=initial_state)
         assert machine.current_state.identifier == initial_state
